--- a/priv/perf/apps/load_test/lib/scenario/deposits.ex
+++ b/priv/perf/apps/load_test/lib/scenario/deposits.ex
@@ -63,7 +63,7 @@ defmodule LoadTest.Scenario.Deposits do
 
         session
         |> Session.add_metric("error_rate", 1)
-        |> Session.add_error(:test, error)
+        |> Session.add_error(:error, error)
     end
   end