--- a/priv/perf/apps/load_test/lib/test_runner/config.ex
+++ b/priv/perf/apps/load_test/lib/test_runner/config.ex
@@ -33,7 +33,8 @@ defmodule LoadTest.TestRunner.Config do
     },
     "utxos" => %{
       token: "0x0000000000000000000000000000000000000000",
-      initial_amount: 760
+      initial_amount: 760,
+      fee: 75
     }
   }