--- a/priv/perf/apps/load_test/lib/watcher_info/balance.ex
+++ b/priv/perf/apps/load_test/lib/watcher_info/balance.ex
@@ -14,7 +14,7 @@
 
 defmodule LoadTest.WatcherInfo.Balance do
   @moduledoc """
-  Functions related to balanes on the childchain
+  Functions related to balances on the childchain
   """
   require Logger