--- a/priv/perf/apps/load_test/lib/watcher_info/utxo.ex
+++ b/priv/perf/apps/load_test/lib/watcher_info/utxo.ex
@@ -64,7 +64,7 @@ defmodule LoadTest.WatcherInfo.Utxo do
     {:error, decoded_response}
   end
 
-  defp find_utxo(decoded_response, _utxo) do
+  defp find_utxo(decoded_response, utxo) do
     do_find_utxo(decoded_response, utxo)
   end