This PR adds utxo perf test that can be run from the command line using `mix run`:

A single iteration of this test consists of the following steps:

1.1 Two accounts are created - the sender and the receiver

2.1 The sender account is funded with `initial_amount`
2.2 The balance on the childchain of the sender is validated using `WatcherInfoAPI.Api.Account.account_get_balance` API.
2.3 Utxos of the sender are validated using `WatcherInfoAPI.Api.Account.account_get_utxos` API

3.1 The sender sends al his value to the receiver
3.2 The balance on the childchain of the sender is validated
3.3 The balance on the childchain of the receiver is validated
3.4 Utxos of the sender are validated
3.5 Utxos of the receiver are validated


You can run it using:
```sh
mix run -e "LoadTest.TestRunner.run()" -- "utxos" 2 5
```
Parameters:
1. test name - utxos
2. tps - 2 t/s
3. period - 5 seconds

So the command above means "Run utxos test for 5 seconds with 2 tests / second". Keep in mind that using 100 t/s causes elixir-omg to start returning internal server error.

You can modify the following test paramers:
      - token -address of the token. default value is "0x0000000000000000000000000000000000000000",
      - initial_amount - this value used when funding the receiver. default value is 760
      - fee - fee used when transferring value from the sender to the receiver

  You can also modify values for the by providing `TEST_CONFIG_PATH` env variable, it should contain
  the path to json file. For example:
```sh
    TEST_CONFIG_PATH=./my_file mix run -e "LoadTest.TestRunner.run()" -- "utxos" 1 5
```