@InoMurko 2020-10-27
Perhaps a bit more effort on the PR description and the test description. `utxos` is not a very good descriptive name of a test. 
the parameters 1 and 1, also don't ring a bell what they mean, could we paramatrice these so that we know what we're setting?
--num_senders
--num_receivers

do the senders and receivers need to be the same? what happens if they're not? if they need to be the same, why are there two separate paramters?