@boolafish 2020-10-29
Can you link this PR to the issue?? so it can be easier tracked. @ayrat555 

I am also curious on what your approach is. Are you removing existing stuffs like `LoadTest.Runner.UtxosLoad` afterward??