@ayrat555 2020-10-29
> how do you get this?? do you run with docker or existing env? Cause I am quite sure we run TPS 100 several times on our sandbox

`WatcherInfoAPI.Api.Account.account_get_utxos(client, body)` starts returning `server:internal_server_error`