@ayrat555 2020-10-29
Also, I see errors in watch_info logs
```
Server: localhost:7534 (http)
watcher_info    | Request: POST /account.get_utxos
watcher_info    | ** (exit) an exception was raised:
watcher_info    |     ** (DBConnection.ConnectionError) connection not available and request was dropped from queue after 171ms. This means requests are coming in and your connection pool cannot serve them fast enough. You can address this by:
watcher_info    | 
watcher_info    |   1. By tracking down slow queries and making sure they are running fast enough
watcher_info    |   2. Increasing the pool_size (albeit it increases resource consumption)
watcher_info    |   3. Allow requests to wait longer by increasing :queue_target and :queue_interval
watcher_info    | 
watcher_info    | See DBConnection.start_link/2 for more information

```