@ayrat555 2020-10-29
> Are you removing existing stuff like LoadTest.Runner.UtxosLoad afterwards??

I didn't plan to do it yet. as @kevsul mentioned, existing tests have a different purpose?